requirements.txt
uvicorn==0.31.1
aiogram==3.13.1
python-dotenv==1.0.0
pydantic==2.9.2
requests==2.32.3
fitz==1.24.11
sentence-transformers==3.2.0
transformers==4.45.2
torch==2.4.1
accelerate